Moved custom api config logic into settings-utils library
